Move fake-provider and fake-block-tracker into a directory and exclude it from coverage
