+++ jest.config.js
@@ -3,6 +3,8 @@ module.exports = {
     '<rootDir>/app/scripts/constants/error-utils.js',
     '<rootDir>/app/scripts/controllers/network/**/*.js',
     '<rootDir>/app/scripts/controllers/network/**/*.ts',
+    '!<rootDir>/app/scripts/controllers/network/**/test/*.ts',
+    '!<rootDir>/app/scripts/controllers/network/**/provider-api-tests/*.ts',
     '<rootDir>/app/scripts/controllers/permissions/**/*.js',
     '<rootDir>/app/scripts/controllers/sign.ts',
     '<rootDir>/app/scripts/flask/**/*.js',
